requirements.txt

# requirements.txt
python-telegram-bot[job-queue]==20.5 # для работы с телеграм-ботом
requests==2.31.0 # для выполнения HTTP-запросов к бирже
python-dotenv==1.0.0 # для работы с переменными окружения